On insert, use the delta value specified in the paper
The original paper gives the value of delta to use on insert as
floor(2*eps*N).  Another description I was reading gave the second
suggestion.  The second suggestion works assuming that g+delta <
2*eps*N, which is required by the original algorithm.  However, merging
tuples (as is done during value merging, an upcoming patch) causes
g+delta to be larger value, breaking the invariant.  So, switch back to
the original version so we don't break when we start merging adjacent
tuples.