make Less marginally faster
This looks great on paper if you aren't using sort.Sort but are
actually just executing this code, where it's quite noticably
faster. On the other hand it's a cool trick.

Observation: If either of the sign bits is set, we're going to
return (s[i] >= s[j]), which it turns out, is exactly the opposite
of (s[i] < s[j]) for non-NaN, and we're not dealing with NaN
anyway because we're looking at the uint64 representations.

So, if we bitwise-or the values, *then* shift, we get a single
bit to test which represents "either sign bit is set". And we
can then streamline one step further by just doing a boolean
xor, also known as `!=`, on the two tests.

If you do this in C, you can save another instruction by not
needing the != 0, or indeed, the conversion to boolean, since
the result of the right-shift is either 0 or 1, and the result
of the < operator is either 0 or 1, so you can just bitwise-xor
them. Woo. An entire instruction.